:recycle: refactor : 활동 도메인 모듈 내 entity와 enum 분리 및 이동
