+++ src/main/kotlin/picklab/backend/activity/domain/enum/RecruitmentStatus.kt
@@ -1,4 +1,4 @@
-package picklab.backend.activity.domain
+package picklab.backend.activity.domain.enum
 
 enum class RecruitmentStatus(
     val label: String,
