+++ src/main/kotlin/picklab/backend/bookmark/domain/Bookmark.kt
@@ -6,7 +6,7 @@ import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
 import org.hibernate.annotations.Comment
-import picklab.backend.activity.domain.Activity
+import picklab.backend.activity.domain.entity.Activity
 import picklab.backend.common.model.BaseEntity
 import picklab.backend.member.domain.Member
